Clean up and make peer indicator a bit nicer. Less data
