+++ src/Components/Details.css
@@ -54,10 +54,6 @@ h4 {
     padding-right: 0.15em;
 }
 
-.download {
-    font-weight: bolder;
-}
-
 .inline {
     display: inline-block;
 }
